fix: add loading ui where it is needed
