+++ src/lib/assets/lottieFingerPrint/fingerPrintLottie.json
