+++ src/lib/assets/lottieFingerPrint/fingerPrintLottieLight.json
