+++ src/routes/[[lang]]/(auth)/login/reset-password/+page.svelte
@@ -10,6 +10,7 @@
 	import type { Feedback } from '$lib/utils/types';
 	import Header from '$lib/components/molecules/Header.svelte';
 	import CircleCheck from '$lib/assets/CircleCheck.svelte';
+	import Loading from '$lib/components/molecules/Loading.svelte';
 
 	let feedback: Feedback = {};
 	let loading = false;
@@ -45,6 +46,7 @@
 	};
 </script>
 
+<Loading {loading} />
 <Header>Reset password</Header>
 <div class="flex min-h-screen flex-col place-content-between">
 	<d-feedback {...feedback} />
